move serialization and resolver into their own files
